fifth

+++ shop/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render,redirect
 from .models import *
-from django.contrib import messages
 from shop.forms import CustomUserForm
 from django.contrib.auth import authenticate,login,logout
 from django.http import JsonResponse
 import json
+from time import sleep
 # Create your views here.
 def home(request):
     prod = Product.objects.filter(trending=1)
@@ -15,13 +15,11 @@ def register(request):
         form = CustomUserForm(request.POST)
         if form.is_valid():
             form.save()
-            messages.success(request,'Registration successful You can Login Now ..!')
             return redirect('login')
     return render(request , "shop/register.html",{'form':form}) 
 def logout_page(request):
-    if request.user.is_authenticated    :
-        logout(request) 
-        messages.success(request,"You Logged out Successfully..")   
+    if request.user.is_authenticated:
+        logout(request)   
         return redirect('/')
 def login_page(request):
     if request.user.is_authenticated :
@@ -33,10 +31,10 @@ def login_page(request):
             user = authenticate(request,username=name,password=pwd)
             if user is not None:
                 login(request,user)
-                messages.success(request,"You Logged in Successfully...")
+               
                 return redirect('/')
             else:
-                messages.error(request,"Invalid Username or Password...")
+                
                 return redirect('login')    
 
     return render(request,'shop/login.html')     
@@ -48,7 +46,7 @@ def products(request,name):
         prod = Product.objects.filter(category__name=name) 
         return render(request,"shop/products.html",{"prod":prod,"category_name":name})        
     else:
-        messages.warning(request,"No Such Category Found")
+        
         return redirect('collections')
 def product_details(request,cname,pname):
     if(Category.objects.filter(status=0,name=cname)):
@@ -56,20 +54,17 @@ def product_details(request,cname,pname):
             product = Product.objects.filter(status=0,name=pname).first()    
             return render(request,"shop/product_details.html",{"prod":product})
         else :
-            messages.warning(request,"No such product found")
+            
             return redirect('collections')       
     else:
-        messages.warning(request,"No Such Category Found")
+       
         return redirect('collections')
 def add_to_cart(request):
     if request.headers.get('x-requested-with')=='XMLHttpRequest':
         if request.user.is_authenticated:
             data = json.load(request)
-            # print(data['pid'])
             product_id = data['pid']
             product_qty = data['product_qty']
-            # print(data['product_qty'])
-            # print(request.user.id)
             product_status = Product.objects.get(id=product_id)
             if product_status :
                 if Cart.objects.filter(user=request.user.id,product_id=product_id):
@@ -88,10 +83,17 @@ def add_to_cart(request):
 def cart(request):
     cartitems = Cart.objects.filter(user=request.user)
     return render(request,'shop/cart.html',{'carts':cartitems})  
-def remove_cart(request,id):
-    cartitem = Cart.objects.get(id=id)
-    cartitem.delete()
-    return redirect('cart')
+def remove_cart(request):
+   if request.user.is_authenticated: 
+     data = json.load(request)
+     print(data['cartid'])
+     id = data['cartid']
+     cartitem = Cart.objects.get(id=id)
+     cartitem.delete()
+     return JsonResponse({'status':'Product removed successfully'},status=200)
+   else :
+     return JsonResponse({'status':'Login to remove'},status=200)  
+    
 def fav_page(request):
    if request.headers.get('x-requested-with')=='XMLHttpRequest':
     if request.user.is_authenticated:
@@ -103,7 +105,7 @@ def fav_page(request):
           return JsonResponse({'status':'Product Already in Favourite'}, status=200)
          else:
           Favourite.objects.create(user=request.user,product_id=product_id)
-          return JsonResponse({'status':'Product Added to Favourite'}, status=200)
+          return JsonResponse({'status':'Product Added to Favourite successfully..'}, status=200)
     else:
       return JsonResponse({'status':'Login to Add Favourite'}, status=200)
    else:
@@ -111,7 +113,14 @@ def fav_page(request):
 def favview(request):
     favourite = Favourite.objects.filter(user=request.user)
     return render(request,'shop/favourite.html',{'fav':favourite})
-def remove_fav(request,id):
+def remove_fav(request):
+   if request.user.is_authenticated:
+    data = json.load(request)  
+    id = data['fid']  
     favourite = Favourite.objects.get(id=id)
     favourite.delete()
-    return redirect('favview')    
+    return JsonResponse({'status':'Favourite product has been removed successfully'},status=200)
+   else:
+    return JsonResponse({'status':'Login to remove '},status=200)
+   
+
